feat!: move dealfactory out of market (#268)
* move dealfactory out of market

* fix tests

* try change imports

* fix vesting

* WIP

* fix yaml

* add info on subgraph

* ts-client fixes

* Update local.json

* artifacts

* atrifacts

* Update local.json

* Update local.json

* lock files

* deployments to stage and dar

* new deployments

* Update local.json

---------
