+++ integration-tests/package.json
@@ -15,7 +15,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@fluencelabs/deal-ts-clients": "https://npm.fluence.dev/@fluencelabs/deal-ts-clients/-/deal-ts-clients-0.3.1-feat-CHAIN-316-d453d94-1359-1.0.tgz",
+    "@fluencelabs/deal-ts-clients": "file:../ts-client",
     "ethers": "6.7.1",
     "vitest": "^1.2.0"
   },
