+++ src/core/modules/market/Market.sol
@@ -5,12 +5,10 @@ pragma solidity ^0.8.19;
 import "@openzeppelin/contracts-upgradeable/proxy/utils/UUPSUpgradeable.sol";
 import "@openzeppelin/contracts-upgradeable/utils/MulticallUpgradeable.sol";
 import "src/core/interfaces/ICore.sol";
-import "src/deal/interfaces/IDeal.sol";
 import "./Matcher.sol";
-import "./DealFactory.sol";
 import "./interfaces/IMarket.sol";
 
-contract Market is UUPSUpgradeable, MulticallUpgradeable, DealFactory, Matcher, IMarket {
+contract Market is UUPSUpgradeable, MulticallUpgradeable, Matcher, IMarket {
     // ------------------ Initializer ------------------
     constructor(ICore core_) BaseModule(core_) {}
