+++ src/core/modules/market/interfaces/IMarket.sol
@@ -2,12 +2,11 @@
 
 pragma solidity ^0.8.19;
 
-import "./IDealFactory.sol";
 import "./IOffer.sol";
 import "./IMatcher.sol";
 
 /// @title Market contract interface
 /// @dev Market contract is responsible for managing the deals and offers
-interface IMarket is IDealFactory, IOffer, IMatcher {
+interface IMarket is IOffer, IMatcher {
     function initialize() external;
 }
