+++ subgraph/configs/local-networks-config.json
@@ -1,16 +1,20 @@
 {
   "local": {
     "Market": {
-      "address": "0x68B1D87F95878fE05B998F19b66F4baba5De1aed",
-      "startBlock": 9
+      "address": "0x3Aa5ebB10DC797CAC828524e59A333d0A371443c",
+      "startBlock": 0
     },
     "Capacity": {
-      "address": "0x3Aa5ebB10DC797CAC828524e59A333d0A371443c",
-      "startBlock": 9
+      "address": "0xc6e7DF5E7b4f2A278906862b61205850344D4e7d",
+      "startBlock": 0
     },
     "Core": {
       "address": "0x0B306BF915C4d645ff596e518fAf3F9669b97016",
-      "startBlock": 9
+      "startBlock": 0
+    },
+    "DealFactory": {
+      "address": "0x1613beB3B2C4f22Ee086B2b38C1476A3cE7f78E8",
+      "startBlock": 7688
     }
   }
 }
